feat: add a policy `deny_job_container_latest_image`
