+++ docs/policies/007.md
@@ -1,0 +1,29 @@
+# deny_job_container_latest_image
+
+Job's container image tag should not be `latest`.
+
+## Why?
+
+Image tags should be pinned with tag or hash.
+
+## Examples
+
+:x:
+
+```yaml
+jobs:
+  container-test-job:
+    runs-on: ubuntu-latest
+    container:
+      image: node:latest
+```
+
+⭕ 
+
+```yaml
+jobs:
+  container-test-job:
+    runs-on: ubuntu-latest
+    container:
+      image: node:10 # Ideally, hash is best
+```
