+++ pkg/cli/run.go
@@ -41,6 +41,7 @@ func (r *Runner) Run(ctx *cli.Context) error {
 		&DenyReadAllPermissionPolicy{},
 		&DenyWriteAllPermissionPolicy{},
 		&DenyInheritSecretsPolicy{},
+		&DenyJobContainerLatestImagePolicy{},
 	}
 	failed := false
 	for _, filePath := range filePaths {
@@ -78,106 +79,4 @@ func (r *Runner) validateWorkflow(ctx *cli.Context, logE *logrus.Entry, cfg *Con
 type Policy interface {
 	Name() string
 	Apply(ctx context.Context, logE *logrus.Entry, cfg *Config, wf *Workflow) error
-}
-
-type Workflow struct {
-	FilePath    string `yaml:"-"`
-	Jobs        map[string]*Job
-	Env         map[string]string
-	Permissions *Permissions
-}
-
-type Job struct {
-	Permissions *Permissions
-	Env         map[string]string
-	Steps       []interface{}
-	Secrets     *JobSecrets
-}
-
-type Permissions struct {
-	m        map[string]string
-	readAll  bool
-	writeAll bool
-}
-
-func (permissions *Permissions) Permissions() map[string]string {
-	if permissions == nil {
-		return nil
-	}
-	return permissions.m
-}
-
-func (permissions *Permissions) ReadAll() bool {
-	if permissions == nil {
-		return false
-	}
-	return permissions.readAll
-}
-
-func (permissions *Permissions) WriteAll() bool {
-	if permissions == nil {
-		return false
-	}
-	return permissions.writeAll
-}
-
-func (permissions *Permissions) IsNil() bool {
-	if permissions == nil {
-		return true
-	}
-	return permissions.m == nil && !permissions.readAll && !permissions.writeAll
-}
-
-func (permissions *Permissions) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	var val interface{}
-	if err := unmarshal(&val); err != nil {
-		return err
-	}
-	return convPermissions(val, permissions)
-}
-
-func convPermissions(src interface{}, dest *Permissions) error { //nolint:cyclop
-	switch p := src.(type) {
-	case string:
-		switch p {
-		case "read-all":
-			dest.readAll = true
-			return nil
-		case "write-all":
-			dest.writeAll = true
-			return nil
-		default:
-			return logerr.WithFields(errors.New("unknown permissions"), logrus.Fields{ //nolint:wrapcheck
-				"permission": p,
-			})
-		}
-	case map[interface{}]interface{}:
-		m := make(map[string]string, len(p))
-		for k, v := range p {
-			ks, ok := k.(string)
-			if !ok {
-				return errors.New("permissions key must be string")
-			}
-			vs, ok := v.(string)
-			if !ok {
-				return errors.New("permissions value must be string")
-			}
-			m[ks] = vs
-		}
-		dest.m = m
-		return nil
-	case map[string]interface{}:
-		m := make(map[string]string, len(p))
-		for k, v := range p {
-			vs, ok := v.(string)
-			if !ok {
-				return errors.New("permissions value must be string")
-			}
-			m[k] = vs
-		}
-		dest.m = m
-		return nil
-	default:
-		return errors.New("permissions must be map[string]string or 'read-all' or 'write-all'")
-	}
 }
